Add files via upload
What is cybersecurity?
Cybersecurity is all about reducing threats when people are in the process of dealing with technology. It encompasses
the full range of protection against any online risk or vulnerability, which comprises information security assurance
and cyber law enforcement. In other words, cybersecurity is the protection of cyber-space (which includes hardware,
software, networks, and their servers, peripheral devices, data and information, and all other components associated
with technology) and internet-connected systems from both internal as well as external threats and cyber-criminals. It
also comprises sub-branches that are specific to different security measures
Why there is such demand for cybersecurity?
Almost everyone from this generation lives in a world of technology where the internet is connected to
nearly every device around us. For cybercriminals and hackers, this is the golden age to tweak with
technical stuff easily. Still, the general users are not aware of the fact that lack of security and improper use
of technology can drag users and employees to doom. That is why specialized security analysts and
penetration testers are hired to secure the entire system.
This subject has gradually evolved to be a new domain of learning and securing different technology such as
programming, web, network, servers, applications, cloud, and many more areas.